remove bug by handling the stack, increase the readability

+++ chapter2/pi/v6/main.c
@@ -15,7 +15,7 @@ double ofs[] __attribute__ ((aligned(32))) = {0.5, 1.5, 2.5, 3.5};
 void calcPi(void)
 {
 	double x;
-	int i;
+	long long i;
 
 	for (i = 0; i < num_steps; i++) {
 		x = (i + 0.5) * step;
@@ -30,27 +30,27 @@ void calcPi_intrinsic(void)
 	__m128d xmm1 = {step, step};
 	__m128d xmm2 = *((__m128d*)ofs);
 	__m128d xmm3, xmm4;
-	int i;
+	long long i;
 
 	for(i = 0; i < num_steps; i+=2) {
 		// Berechne (i+0.5f)*step
-                xmm4 = __builtin_ia32_mulpd(xmm1, xmm2);
+		xmm4 = __builtin_ia32_mulpd(xmm1, xmm2);
 
 		// Quadriere das Zwischenergebniss
-                // und erhöhe um eins
-                xmm4 = __builtin_ia32_mulpd(xmm4, xmm4);
-                xmm4 = __builtin_ia32_addpd(xmm4, *((__m128d*) one));
+		// und erhöhe um eins
+		xmm4 = __builtin_ia32_mulpd(xmm4, xmm4);
+		xmm4 = __builtin_ia32_addpd(xmm4, *((__m128d*) one));
 
 		// teile 4 durch das Zwischenergebnis
-                xmm3 = *((__m128d*) four);
-                xmm3 = __builtin_ia32_divpd(xmm3, xmm4);
+		xmm3 = *((__m128d*) four);
+		xmm3 = __builtin_ia32_divpd(xmm3, xmm4);
 
-                // Summiere die ermittelten Rechteckshöhen auf
-                xmm0 = __builtin_ia32_addpd(xmm0, xmm3);
+		// Summiere die ermittelten Rechteckshöhen auf
+		xmm0 = __builtin_ia32_addpd(xmm0, xmm3);
 
-                //Laufzäler erhöhen und
-                //zum Schleifenanfang springen
-                xmm2 = __builtin_ia32_addpd(xmm2, *((__m128d*) two));
+		// Laufzäler erhöhen und
+		// zum Schleifenanfang springen
+		xmm2 = __builtin_ia32_addpd(xmm2, *((__m128d*) two));
 	}
 
 	sum = xmm0[0] + xmm0[1];
